Добавлена лента новостей на все методы

+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -7,8 +7,7 @@ import org.springframework.util.StringUtils;
 import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.User;
-import ru.yandex.practicum.filmorate.model.feed.EventTypeEnum;
-import ru.yandex.practicum.filmorate.model.feed.OperationEnum;
+import ru.yandex.practicum.filmorate.model.Feed;
 import ru.yandex.practicum.filmorate.storage.user.UserStorage;
 
 import java.time.LocalDate;
@@ -18,12 +17,10 @@ import java.util.List;
 @Slf4j
 public class UserService {
     private final UserStorage userStorage;
-    private final FeedService feedService;
 
     @Autowired
-    public UserService(UserStorage userStorage, FeedService feedService) {
+    public UserService(UserStorage userStorage) {
         this.userStorage = userStorage;
-        this.feedService = feedService;
     }
 
     public User getUser(Integer id) {
@@ -53,17 +50,17 @@ public class UserService {
 
     public User addFriend(Integer id, Integer friendId) {
         log.info(String.format("UserService: Добавление пользователем %d друга %d", id, friendId));
-
-        feedService.toFeed(friendId, id, EventTypeEnum.FRIEND, OperationEnum.ADD);
-
+        if (id < 0 || friendId < 0) {
+            throw new NotFoundException("Идентификаторы не могут быть отрицательными.");
+        }
         return userStorage.addFriend(id, friendId);
     }
 
     public User deleteFriend(Integer id, Integer friendId) {
         log.info(String.format("UserService: Удаление пользователем %d друга %d", id, friendId));
-
-        feedService.toFeed(friendId, id, EventTypeEnum.FRIEND, OperationEnum.REMOVE);
-
+        if (id < 0 || friendId < 0) {
+            throw new ValidationException("Идентификаторы не могут быть отрицательными.");
+        }
         return userStorage.deleteFriend(id, friendId);
     }
 
@@ -73,8 +70,20 @@ public class UserService {
     }
 
     public List<User> getCommonFriends(Integer id, Integer otherId) {
+        if (id < 0 || otherId < 0) {
+            throw new ValidationException("Идентификаторы не могут быть отрицательными.");
+        }
         log.info(String.format("UserService: Получение списка друзей пользователя %d, общих с пользователем %d", id, otherId));
         return userStorage.getCommonFriends(id, otherId);
+    }
+
+    public List<Feed> getUserFeed(Integer userId) {
+        if (isExist(userId)) {
+            log.info("Новостная лента пользователя: {} {}", userStorage.getUser(userId).getName(), userStorage.getUserFeed(userId));
+            return userStorage.getUserFeed(userId);
+        } else {
+            throw new NotFoundException("Такого пользователя не существует.");
+        }
     }
 
     public void validateUser(User user) {
@@ -105,5 +114,14 @@ public class UserService {
         }
         if (!StringUtils.hasText(user.getName()))
             user.setName(user.getLogin());
+    }
+
+    private boolean isExist(int id) {
+        for (User user : userStorage.getAllUsers()) {
+            if (id == user.getId()) {
+                return true;
+            }
+        }
+        return false;
     }
 }
